Clarify root package acceptance tests
- Split process, HTTP, documented-example, and canonical distribution acceptance into explicit targets.
- Remove root scenarios already owned by geam-cli and the Provider SDK.
- Prepare only the dependencies required by each distribution boundary.
