+++ tests/cross_crate_http/service.rs
